db: consult the seccomp(2) syscall when validating unknown actions
Split actions into two different classes by putting all the actions that
existed prior to the inception of libseccomp into an "assumed" class and
then consult the new SECCOMP_GET_ACTION_AVAIL seccomp() operation for
all others.

This allows for libseccomp to work with actions which are unknown to it
but supported by the current running kernel.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>