all: add support for new log filter flag
Extend libseccomp to support SECCOMP_FILTER_FLAG_LOG, which is intended
to cause log events for all actions taken by a filter except for
SCMP_ACT_ALLOW actions. This is done via a new filter attribute called
SCMP_FLTATR_CTL_LOG that is off by default.

The 13-basic-attrs test is updated to set and, only if the flag is
supported by the kernel, verify that the flag was successfully set.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>