all: add seccomp_action_valid(3) for validating actions
Add a new function, which is exported publicly, to allow applications to
check if a given action is supported. This change essentially allows
db_action_valid() to be used by applications building against
libseccomp. It is most useful for checking to see if an action is valid
before calling seccomp_init(3) when attempting to use an action that was
added to the Linux kernel after libseccomp's inception.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>