all: add global attribute to skip kernel validation of actions
This function introduces two new concepts to libseccomp. The first is
the idea of global attributes which are not tied to a specific filter.
The second is a global attribute, SCMP_GLBATR_CTL_KCHECKACTS, which
indicates that libseccomp should not consult the kernel when validating
actions which were added to the Linux kernel after libseccomp's
inception.

The global attribute concept is nice in that global attributes can be
set prior to calling seccomp_init(3) but it requires a slight abuse of
the existing seccomp_attr_set(3)/seccomp_attr_get(3) APIs since the
attribute must fit inside of the existing scmp_filter_attr enum. An
attempt is made to draw the distinction between filter attrs and global
attrs in the naming of the attribute and in the seccomp_attr_set(3) man
page.

The get_attr() and set_attr() functions in the cython wrapper are
essentially copied to have a global function that doesn't require a
SyscallFilter object so that we know when NULL should be passed to the
corresponding libseccomp functions.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>